\name{mumps}
\alias{mumps}
\docType{data}
\title{Mumps}
\description{Monthly registered cases of mumps in New York City, January 1928
through June 1972.
}
\usage{data(mumps)}
\format{
534 observations 
  \describe{
    \item{mumps}{a numeric vector of registred cases.}
  }
}
\references{
 Hipel KW, McLeod IA (1994). Time Series Modeling of Water Resources
  and Environmental Systems. Elsevier Science Publishers B.V. (North-Holland).
  }
\keyword{datasets}
