\name{mumpsdat}
\alias{mumpsdat}
\docType{data}
\title{Mumps}
\description{Monthly registered cases of mumps in New York City, January 1928
through June 1972.
}
\usage{data(mumpsdat)}
\format{
  A data frame with 534 observations on the following 4 variables.
  \describe{
    \item{Year}{a numeric vector containing the year of the observations.}
    \item{month}{a factor taking values "Jan" through "Dec".}
    \item{mo}{a numeric vector with the month recorded as 1 through 12.}
    \item{mumps}{a numeric vector of registred cases.}
  }
}
\references{
 Hipel KW, McLeod IA (1994). Time Series Modeling of Water Resources
  and Environmental Systems. Elsevier Science Publishers B.V. (North-Holland).
  }
\keyword{datasets}
