\name{vandrivers}
\alias{vandrivers}
\docType{data}
\title{Vandrivers}
\description{
The monthly numbers of light goods van drivers killed in
road accidents, from January 1969 to December 1984}
\usage{data(vandrivers)}
\format{
  A data frame with 192 observations on the following variables.
  \describe{
    \item{y}{a numeric vector giving the number of deaths.}
    \item{seatbelt}{a numeric vector with values 0 before February 1983
      and 1 after. This indicates when the seat belt legislation law was
    introduced.}
  }
}
\keyword{datasets}
